product

+++ products/views.py
@@ -8,17 +8,59 @@ from .serializers import ProductSerializer
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.decorators import permission_classes
+from django.core.paginator import Paginator
+from rest_framework.pagination import PageNumberPagination
+
+
+class ProductPagination(PageNumberPagination):
+    page_query_param = "page"
 
 
 class ProductListAPIView(APIView):
-    # permission_classes = [IsAuthenticated]
+    pagination_class = ProductPagination
+    serializer_class = ProductSerializer
 
+    @property
+    def paginator(self):
+        if not hasattr(self, "_paginator"):
+            if self.pagination_class is None:
+                self._paginator = None
+            else:
+                self._paginator = self.pagination_class()
+        else:
+            pass
+        return self._paginator
+
+    def paginate_queryset(self, queryset):
+        if self.paginator is None:
+            return None
+        return self.paginator.paginate_queryset(queryset, self.request, view=self)
+
+    def get_paginated_response(self, data):
+        assert self.paginator is not None
+        return self.paginator.get_paginated_response(data)
+
+    # 목록
     def get(self, request):
         Products = Product.objects.all()
-        serializer = ProductSerializer(Products, many=True)
-        return Response(serializer.data)
+        page = self.paginate_queryset(Products)
+        print(page)
+        if page is not None:
+            serializer = self.get_paginated_response(
+                self.serializer_class(page, many=True).data
+            )
+        else:
+            serializer = self.serializer_class(Products, many=True)
 
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    # 등록
+    @permission_classes([IsAuthenticated])
     def post(self, request):
+        user = request.user
+        print(user.id)
+        request.data["author"] = user.id
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -26,7 +68,6 @@ class ProductListAPIView(APIView):
 
 
 class ProductDetailAPIView(APIView):
-    # 두 번 이상 반복되는 로직은 함수로 빼면 좋습니다👀
     def get_object(self, pk):
         return get_object_or_404(Product, pk=pk)
 
@@ -35,57 +76,33 @@ class ProductDetailAPIView(APIView):
         serializer = ProductSerializer(Product)
         return Response(serializer.data)
 
+    # 수정
+    @permission_classes([IsAuthenticated])
     def put(self, request, pk):
+        user = request.user
         Product = self.get_object(pk)
-        serializer = ProductSerializer(Product, data=request.data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if user.id == Product.author_id:
+            serializer = ProductSerializer(Product, data=request.data, partial=True)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response(
+                {"message": "Author only."}, status=status.HTTP_400_BAD_REQUEST
+            )
 
+        return Response({"message": "Bad Request."}, status=status.HTTP_400_BAD_REQUEST)
+
+    # 삭제
+    @permission_classes([IsAuthenticated])
     def delete(self, request, pk):
+        user = request.user
         Product = self.get_object(pk)
-        Product.delete()
-        data = {"pk": f"{pk} is deleted."}
-        return Response(data, status=status.HTTP_200_OK)
-
-
-def Product_list_html(request):
-    Products = Product.objects.all()
-    context = {"Products": Products}
-    return render(request, "Products/Products.html", context)
-
-
-@api_view(["GET", "POST"])
-def Product_list(request):
-    if request.method == "GET":
-        Products = Product.objects.all()
-        serializer = ProductSerializer(Products, many=True)
-        return Response(serializer.data)
-    elif request.method == "POST":
-        serializer = ProductSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(["GET", "PUT", "DELETE"])
-def Product_detail(request, pk):
-    Product = get_object_or_404(Product, pk=pk)
-    print(request.method)
-    if request.method == "GET":
-        serializer = ProductSerializer(Product)
-        print("get")
-        return Response(serializer.data)
-    elif request.method == "PUT":
-        print("put")
-        serializer = ProductSerializer(Product, data=request.data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    elif request.method == "DELETE":
-        print("delete")
-        Product.delete()
-        data = {"delete": f"Product({pk}) is deleted."}
-        return Response(data, status=status.HTTP_200_OK)
+        if user.id == Product.author_id:
+            Product.delete()
+            data = {"pk": f"{pk} is deleted."}
+            return Response(data, status=status.HTTP_200_OK)
+        else:
+            return Response(
+                {"message": "Author only."}, status=status.HTTP_400_BAD_REQUEST
+            )
